Refactor SliderControl into own file, migrate to TS
