+++ src/components/Camera/AICamera/AIDebugButton.js
@@ -1,16 +1,11 @@
-import Slider from "@react-native-community/slider";
 import classnames from "classnames";
 import {
-  Body1,
-  Heading4,
-  INatIconButton,
-  P
+  INatIconButton
 } from "components/SharedComponents";
 import {
   CIRCLE_SIZE
 } from "components/SharedComponents/Buttons/TransparentCircleButton.tsx";
 import { View } from "components/styledComponents";
-import { round } from "lodash";
 import React, { useState } from "react";
 import {
   Modal,
@@ -19,37 +14,7 @@ import {
 import { useDebugMode } from "sharedHooks";
 import colors from "styles/tailwindColors";
 
-const SLIDER_STYLE = { display: "flex", flexGrow: 1, height: 44 };
-
-const SliderControl = ( {
-  name,
-  value,
-  setValue,
-  min,
-  max,
-  precision = 0,
-  step = 1
-} ) => (
-  <P>
-    {/* eslint-disable-next-line i18next/no-literal-string */}
-    <Heading4 className="text-white">{ `${name} (${min}-${max})` }</Heading4>
-    <View className="flex-row items-center h-fit">
-      <Body1 className="w-10 m-3 text-white">{round( value, precision )}</Body1>
-      <Slider
-        style={SLIDER_STYLE}
-        minimumValue={min}
-        maximumValue={max}
-        minimumTrackTintColor="#FFFFFF"
-        maximumTrackTintColor={colors.black}
-        renderStepNumber
-        tapToSeek
-        step={step}
-        value={value}
-        onValueChange={changedValue => setValue( changedValue )}
-      />
-    </View>
-  </P>
-);
+import SliderControl from "./SliderControl";
 
 const AIDebugButton = ( {
   confidenceThreshold,
